+++ CHANGELOG.md
@@ -42,6 +42,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [#303](https://github.com/babylonlabs-io/vigilante/pull/303) chore: sew reduce logs
 * [#308](https://github.com/babylonlabs-io/vigilante/pull/308) chore: reporter handle duplicate submissions
 
+### Bug Fixes
+
+* [#311](https://github.com/babylonlabs-io/vigilante/pull/311) add yaml metadata to cfg
+
 ## v0.23.2
 
 ### Improvements